Moved callable statement related handler into dedicated package
