+++ server/src/main/java/com/consol/citrus/db/server/JdbcServer.java
@@ -21,9 +21,9 @@ import com.consol.citrus.db.server.controller.JdbcController;
 import com.consol.citrus.db.server.controller.RuleBasedController;
 import com.consol.citrus.db.server.controller.SimpleJdbcController;
 import com.consol.citrus.db.server.exceptionhandler.JdbcServerExceptionHandler;
-import com.consol.citrus.db.server.handler.CloseCallableStatementHandler;
-import com.consol.citrus.db.server.handler.CreateCallableStatementHandler;
-import com.consol.citrus.db.server.handler.ExecuteCallableStatementHandler;
+import com.consol.citrus.db.server.handler.callablestatement.CloseCallableStatementHandler;
+import com.consol.citrus.db.server.handler.callablestatement.CreateCallableStatementHandler;
+import com.consol.citrus.db.server.handler.callablestatement.ExecuteCallableStatementHandler;
 import com.consol.citrus.db.server.handler.connection.CloseConnectionHandler;
 import com.consol.citrus.db.server.handler.connection.CommitTransactionStatementsHandler;
 import com.consol.citrus.db.server.handler.connection.GetTransactionStateHandler;
